Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main(){
     int y;
     int z;
     scanf("%d %d %d",&x,&y,&z);
-    float a=(x+y+z)/3;
+    int a=(x+y+z)/3;
     printf("Average: %.2f",a);
     return 0;
 }
